requirements.txt
alembic==1.4.2
boto3==1.17.22
botocore==1.20.22
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
DateTime==4.3
decorator==4.4.2
Flask==1.1.2
Flask-JSGlue==0.3.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.8
Jinja2==2.11.3
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2021.1
regex==2020.5.14
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.16
urllib3==1.26.3
Werkzeug==0.16.0
WTForms==2.2.1
zope.interface==5.2.0
itsdangerous==2.0.1